Restructure the folders
