+++ .gitignore
@@ -107,6 +107,9 @@ _UpgradeReport_Files/
 Backup*/
 UpgradeLog*.XML
 
+# Don't take special visual studio folder
+.vs/
+
 # Don't take final configuration file
 plugin.json
 newrelic.json
