+++ ScalableBytes.NewRelic.AzureStorageQueueSize.Plugin.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
 VisualStudioVersion = 12.0.40629.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ScalableBytes.NewRelic.AzureStorageQueueSize.Plugin", "ScalableBytes.NewRelic.AzureStorageQueueSize.Plugin.csproj", "{FD134883-5BBF-4B95-9345-1B521B8D896F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ScalableBytes.NewRelic.AzureStorageQueueSize.Plugin", "ScalableBytes.NewRelic.AzureStorageQueueSize.Plugin\ScalableBytes.NewRelic.AzureStorageQueueSize.Plugin.csproj", "{FD134883-5BBF-4B95-9345-1B521B8D896F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
