+++ ScalableBytes.NewRelic.AzureStorageQueueSize.Plugin/AgentManager.cs
