+++ ScalableBytes.NewRelic.AzureStorageQueueSize.Plugin/App.config
