+++ ScalableBytes.NewRelic.AzureStorageQueueSize.Plugin/Models/SystemConfiguration.cs
