+++ ScalableBytes.NewRelic.AzureStorageQueueSize.Plugin/Properties/AssemblyInfo.cs
