+++ ScalableBytes.NewRelic.AzureStorageQueueSize.Plugin/config/newrelic.template.json
