+++ ScalableBytes.NewRelic.AzureStorageQueueSize.Plugin/config/plugin.template.json
