+++ ScalableBytes.NewRelic.AzureStorageQueueSize.Plugin/packages.config
